Prepare for Release

+++ lib/Entity/Media.php
@@ -37,6 +37,7 @@ use Xibo\Factory\WidgetFactory;
 use Xibo\Helper\Config;
 use Xibo\Helper\Date;
 use Xibo\Helper\Log;
+use Xibo\Helper\Sanitize;
 use Xibo\Storage\PDOConnect;
 
 /**
@@ -204,6 +205,15 @@ class Media implements \JsonSerializable
     }
 
     /**
+     * Sets the Owner
+     * @param int $ownerId
+     */
+    public function setOwner($ownerId)
+    {
+        $this->ownerId = $ownerId;
+    }
+
+    /**
      * Validate
      * @param array $options
      */
@@ -337,8 +347,8 @@ class Media implements \JsonSerializable
             if ($parentMedia != null)
                 $widget->assignMedia($parentMedia->mediaId);
 
-            // This action might result in us deleting a widget
-            if (count($widget->mediaIds) <= 0)
+            // This action might result in us deleting a widget (unless we are a temporary file with an expiry date)
+            if ($this->expires == 0 && count($widget->mediaIds) <= 0)
                 $widget->delete();
             else
                 $widget->save(['saveWidgetOptions' => false]);
@@ -394,10 +404,11 @@ class Media implements \JsonSerializable
         $this->saveFile();
 
         // Update the MD5 and storedAs to suit
-        PDOConnect::update('UPDATE `media` SET md5 = :md5, fileSize = :fileSize, storedAs = :storedAs WHERE mediaId = :mediaId', [
+        PDOConnect::update('UPDATE `media` SET md5 = :md5, fileSize = :fileSize, storedAs = :storedAs, duration = :duration WHERE mediaId = :mediaId', [
             'fileSize' => $this->fileSize,
             'md5' => $this->md5,
             'storedAs' => $this->storedAs,
+            'duration' => $this->duration,
             'mediaId' => $this->mediaId
         ]);
     }
@@ -425,7 +436,8 @@ class Media implements \JsonSerializable
                 expires = :expires,
                 moduleSystemFile = :moduleSystemFile,
                 editedMediaId = :editedMediaId,
-                isEdited = :isEdited
+                isEdited = :isEdited,
+                userId = :userId
            WHERE mediaId = :mediaId
         ', [
             'name' => $this->name,
@@ -437,6 +449,7 @@ class Media implements \JsonSerializable
             'moduleSystemFile' => $this->moduleSystemFile,
             'editedMediaId' => $this->parentId,
             'isEdited' => $this->isEdited,
+            'userId' => $this->ownerId,
             'mediaId' => $this->mediaId
         ]);
     }
@@ -490,6 +503,9 @@ class Media implements \JsonSerializable
         // Work out the MD5
         $this->md5 = md5_file($libraryFolder . $this->storedAs);
         $this->fileSize = filesize($libraryFolder . $this->storedAs);
+
+        // Set the duration
+        $this->duration = $this->determineDuration();
     }
 
     /**
@@ -545,7 +561,7 @@ class Media implements \JsonSerializable
 
         try {
             $client = new Client();
-            $client->get($this->fileName, Config::getProxy(['save_to' => $fileHandle]));
+            $client->get($this->fileName, Config::getGuzzelProxy(['save_to' => $fileHandle]));
         }
         catch (RequestException $e) {
             Log::error('Unable to get %s, %s', $this->fileName, $e->getMessage());
@@ -553,5 +569,38 @@ class Media implements \JsonSerializable
 
         // Change the filename to our temporary file
         $this->fileName = $storedAs;
+    }
+
+    /**
+     * Determine the duration of this media based on its type
+     *  - perhaps we should instantiate a WidgetModule for this?
+     */
+    public function determineDuration()
+    {
+        switch ($this->mediaType) {
+
+            case 'image':
+                $duration = Config::GetSetting('jpg_length');
+                break;
+
+            case 'flash':
+                $duration = Config::GetSetting('swf_length');
+                break;
+
+            case 'powerpoint':
+                $duration = Config::GetSetting('ppt_length');
+                break;
+
+            case 'video':
+                $info = new \getID3();
+                $file = $info->analyze(Config::GetSetting('LIBRARY_LOCATION') . $this->storedAs);
+                $duration = intval(Sanitize::getDouble('playtime_seconds', 0, $file));
+                break;
+
+            default:
+                $duration = 0;
+        }
+
+        return $duration;
     }
 }
